requirements.txt
torch
torchvision
torchaudio
git+https://github.com/Genesis-Embodied-AI/Genesis.git
rsl-rl-lib==2.3.3absl-py==2.3.0
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
coacd==1.0.7
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
etils==1.12.2
filelock==3.18.0
fonttools==4.58.4
freetype-py==2.5.1
fsspec==2025.5.1
genesis-world @ git+https://github.com/Genesis-Embodied-AI/Genesis.git@124913164184114c3f712d16e6b0511951c67410
gitdb==4.0.12
GitPython==3.1.44
glfw==2.9.0
grpcio==1.73.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.8
lazy_loader==0.4
libigl==2.6.1
llvmlite==0.44.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
moviepy==2.2.1
mpmath==1.3.0
mujoco==3.3.3
mypy_extensions==1.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
ompl==1.7.0
onnx==1.18.0
opencv-python==4.11.0.86
OpenEXR==3.3.4
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
plotly==5.24.1
pooch==1.8.2
proglog==0.1.12
protobuf==6.31.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycollada==0.9.1
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
PyGEL3D==0.6.1
pyglet==2.1.6
pygltflib==1.16.0
Pygments==2.19.2
pymeshlab==2023.12.post3
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyvista==0.45.2
requests==2.32.4
rich==14.0.0
rsl-rl-lib==2.3.3
rtree==1.4.0
scikit-image==0.25.2
scipy==1.15.3
scooby==0.10.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
taichi==1.7.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tetgen==0.6.5
tifffile==2025.5.10
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
trimesh==4.6.13
triton==3.3.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
usd-core==25.5.1
vtk==9.4.2
Werkzeug==3.1.3
wrapt==1.17.2
z3-solver==4.15.1.0
zipp==3.23.0
